+++ Modules/Microsoft365DSC/DSCResources/MSFT_EXOMobileDeviceMailboxPolicy/MSFT_EXOMobileDeviceMailboxPolicy.psm1
@@ -350,11 +350,13 @@ function Get-TargetResource
                 MaxEmailHTMLBodyTruncationSize           = $MobileDeviceMailboxPolicy.MaxEmailHTMLBodyTruncationSize
                 MaxInactivityTimeLock                    = $MobileDeviceMailboxPolicy.MaxInactivityTimeLock
                 MaxPasswordFailedAttempts                = $MobileDeviceMailboxPolicy.MaxPasswordFailedAttempts
-                MinPasswordComplexCharacters             = $MobileDeviceMailboxPolicy.MinPasswordComplexCharacters
+                # The MinPasswordComplexCharacters property is an integer, but the DSC resource expects a string.
+                MinPasswordComplexCharacters             = $MobileDeviceMailboxPolicy.MinPasswordComplexCharacters.ToString()
                 MinPasswordLength                        = $MobileDeviceMailboxPolicy.MinPasswordLength
                 PasswordEnabled                          = $MobileDeviceMailboxPolicy.PasswordEnabled
                 PasswordExpiration                       = $MobileDeviceMailboxPolicy.PasswordExpiration
-                PasswordHistory                          = $MobileDeviceMailboxPolicy.PasswordHistory
+                # The PasswordHistory property is an integer, but the DSC resource expects a string.
+                PasswordHistory                          = $MobileDeviceMailboxPolicy.PasswordHistory.ToString()
                 PasswordRecoveryEnabled                  = $MobileDeviceMailboxPolicy.PasswordRecoveryEnabled
                 RequireDeviceEncryption                  = $MobileDeviceMailboxPolicy.RequireDeviceEncryption
                 RequireEncryptedSMIMEMessages            = $MobileDeviceMailboxPolicy.RequireSignedSMIMEMessages